Add first version of URL shortening algorithm
We're using the SHA1 algorithm to generate a hash for the URL without
its scheme. That way we can take into account anything else that arrives
in valid URLs without worrying too much about details.

I picked SHA1 over MD5 because Git uses SHA1 to generate the commit
hashes, and it also uses a short notation for them. It allows you to
fetch a given commit by its first 8 characters. This is quite close to
the use case we have here.