Redirect client to original URL on short URL
When the application receives a valid short URL that it can find in the
database, it will redirect the client to the original URL.

Along with that, it will also increment the hit count for the short URL
record in the database. That way we can keep track of which URLs are
more accessed and we can also clean up old URLs later on.

One tradeoff in this approach is that the `links` table will grow over
time and it will get slower both to read and write in as time passes.
We need a way to clean that table and remove older entries.

Another solution would be to use an in-memory database instead. For
Elixir that would mean ETS, but the technical assessment requires us to
use a persistent database.